requirements.txt
altair
numpy
pandas
pydeck
streamlit
streamlit-authenticator
streamlit_extras
google-cloud-bigquery
google-auth 
google-auth-httplib2 
google-api-python-client
openai
google-auth-oauthlib
google-api-client